Fix/improve the "share" SVG icon
The "share" button SVG icon does not look correct. The arrow head is
split. This is a common issue with arrows, as the head is made of three
strokes joining at one point.

SVG path spec allows some optimization for this. The current path data
uses the `M`/`m` command to denote each segment as a separate stroke.
This commit fixes the "split head" problem by drawing a V-shaped stroke
at the head, instead of separate segments. I also tried to optimize the
path data as much as possible, while keeping the identical render
result.

In addition, the path attributes `strokeLinecap="round"` and
`strokeLinejoin="round"` are not having any effect at all (I checked
with Firefox and Chromium). It is possible to fix this, but in my
opinion, they bloat the payload without much aesthetic improvement. So
this commit is my suggestion that we just remove them.